+++ server.js
@@ -2,57 +2,57 @@ const express = require("express");
 require('dotenv').config();
 const bodyParser = require("body-parser");
 const CORS = require("cors");
-const movies = require('./data/moviedb')
+// const movies = require('./data/moviedb')
 
 const app = express();
 
 app.use(bodyParser.json());
 app.use(CORS());
 
-// let movies = [
-//   {
-//     id: 0,
-//     title: "The Godfather",
-//     director: "Francis Ford Coppola",
-//     metascore: 100,
-//     stars: ["Marlon Brando", "Al Pacino", "Robert Duvall"]
-//   },
-//   {
-//     id: 1,
-//     title: "Star Wars",
-//     director: "George Lucas",
-//     metascore: 92,
-//     stars: ["Mark Hamill", "Harrison Ford", "Carrie Fisher"]
-//   },
-//   {
-//     id: 2,
-//     title: "The Lord of the Rings: The Fellowship of the Ring",
-//     director: "Peter Jackson",
-//     metascore: 92,
-//     stars: ["Elijah Wood", "Ian McKellen", "Orlando Bloom"]
-//   },
-//   {
-//     id: 3,
-//     title: "Terminator 2: Judgement Day",
-//     director: "James Cameron",
-//     metascore: 94,
-//     stars: ["Arnold Schwarzenegger", "Edward Furlong", "Linda Hamilton"]
-//   },
-//   {
-//     id: 4,
-//     title: "Dumb and Dumber",
-//     director: "The Farely Brothers",
-//     metascore: 76,
-//     stars: ["Jim Carrey", "Jeff Daniels", "Lauren Holly"]
-//   },
-//   {
-//     id: 5,
-//     title: "Tombstone",
-//     director: "George P. Cosmatos",
-//     metascore: 89,
-//     stars: ["Kurt Russell", "Bill Paxton", "Sam Elliot"]
-//   }
-// ];
+let movies = [
+  {
+    id: 0,
+    title: "The Godfather",
+    director: "Francis Ford Coppola",
+    metascore: 100,
+    stars: ["Marlon Brando", "Al Pacino", "Robert Duvall"]
+  },
+  {
+    id: 1,
+    title: "Star Wars",
+    director: "George Lucas",
+    metascore: 92,
+    stars: ["Mark Hamill", "Harrison Ford", "Carrie Fisher"]
+  },
+  {
+    id: 2,
+    title: "The Lord of the Rings: The Fellowship of the Ring",
+    director: "Peter Jackson",
+    metascore: 92,
+    stars: ["Elijah Wood", "Ian McKellen", "Orlando Bloom"]
+  },
+  {
+    id: 3,
+    title: "Terminator 2: Judgement Day",
+    director: "James Cameron",
+    metascore: 94,
+    stars: ["Arnold Schwarzenegger", "Edward Furlong", "Linda Hamilton"]
+  },
+  {
+    id: 4,
+    title: "Dumb and Dumber",
+    director: "The Farely Brothers",
+    metascore: 76,
+    stars: ["Jim Carrey", "Jeff Daniels", "Lauren Holly"]
+  },
+  {
+    id: 5,
+    title: "Tombstone",
+    director: "George P. Cosmatos",
+    metascore: 89,
+    stars: ["Kurt Russell", "Bill Paxton", "Sam Elliot"]
+  }
+];
 
 let movieId = movies.length;
